requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai 
google-generativeai
langchain-experimental
langgraph
pandas
matplotlib
python-dotenv
streamlit
tabulate
openpyxl
pyarrow
langchain 
langchain-core 
langgraph 
langchain-community
langchain-google-genai